test(agent-bundle): run pack-and-install suites through test:packed instead of the default integration pool
Every npm pack + clean-install suite (dev-workbench-packaging,
packed-consumer, packed-native-smoke, release-audit,
rsc-runtime-optional-packaging) leaves the serialized integration pool
and moves to a dedicated packedTestFiles list that the unit pool also
excludes. The three pack+install cases in public-api.test.ts split into
public-api-packed.test.ts so the cheap export and built-entrypoint
checks stay in the default loop. test:packed now lists the split file
plus rsc-runtime-optional-packaging and packed-native-smoke, keeping
per-PR CI coverage through the release-gates job's check:release run.
